Explicitly set sidekiq environment
Sidekiq looks at the RAILS_ENV and APP_ENV environment variables and at
some point started prefering APP_ENV over RAILS_ENV. In our case, this
broke the deployment when we updated sidekiq, since the APP_ENV can in
some cases be different from the RAILS_ENV, leading to sidekiq looking
for an environment that wasn't there, but without complaining about it,
but just running with all the Rails defaults. Since this is a Rails
image, I think it should be safe to just set the sidekiq environment to
RAILS_ENV, since they should be the same anyway.